Move custom hostfxr loading logic to installation library
